+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 esma_set_this ()
 
-option(BUILD_GEOS_GTFV3_INTERFACE "Build GEOS-gtFV3 interface" OFF)
+option(BUILD_PYFV3_INTERFACE "Build pyFV3 interface" OFF)
 
 set (srcs
   sw.f90 jw.f90 testcases_3_4_5_6_stand_alone.f90
@@ -22,17 +22,17 @@ set (srcs
   fv_regridding_utils.F90
   rs_scaleMod.F90
   )
-if (BUILD_GEOS_GTFV3_INTERFACE)
+if (BUILD_PYFV3_INTERFACE)
   list (APPEND srcs
-  python/interface/geos_gtfv3_interface.f90
-  python/interface/geos_gtfv3_interface.c)
+  python/interface/cffi_lib/interface.f90
+  python/interface/cffi_lib/interface.c)
 endif ()
 
-if (BUILD_GEOS_GTFV3_INTERFACE)
+if (BUILD_PYFV3_INTERFACE)
 
-  message(STATUS "Building GEOS-gtFV3 interface")
+  message(STATUS "Building pyFV3 interface")
 
-  add_definitions(-DRUN_GTFV3)
+  add_definitions(-DRUN_PYFV3)
   
   # The Python library creation requires mpiexec/mpirun to run on a
   # compute node. Probably a weird SLURM thing?
@@ -40,28 +40,28 @@ if (BUILD_GEOS_GTFV3_INTERFACE)
   find_package(Python3 COMPONENTS Interpreter REQUIRED)
 
   # Set up some variables in case names change
-  set(GEOS_GTFV3_INTERFACE_LIBRARY     ${CMAKE_CURRENT_BINARY_DIR}/libgeos_gtfv3_interface_py.so)
-  set(GEOS_GTFV3_INTERFACE_HEADER_FILE ${CMAKE_CURRENT_BINARY_DIR}/geos_gtfv3_interface_py.h)
-  set(GEOS_GTFV3_INTERFACE_FLAG_HEADER_FILE ${CMAKE_CURRENT_SOURCE_DIR}/python/interface/fv_flags.h)
-  set(GEOS_GTFV3_INTERFACE_SRCS        ${CMAKE_CURRENT_SOURCE_DIR}/python/interface/geos_gtfv3_interface.py)
+  set(PYFV3_INTERFACE_LIBRARY     ${CMAKE_CURRENT_BINARY_DIR}/libpyFV3_interface_py.so)
+  set(PYFV3_INTERFACE_HEADER_FILE ${CMAKE_CURRENT_BINARY_DIR}/pyfv3_interface_py.h)
+  set(PYFV3_INTERFACE_FLAG_HEADER_FILE ${CMAKE_CURRENT_SOURCE_DIR}/python/interface/cffi_lib/fv_flags.h)
+  set(PYFV3_INTERFACE_SRCS        ${CMAKE_CURRENT_SOURCE_DIR}/python/interface/cffi_lib/interface.py)
 
   # This command creates the shared object library from Python
   add_custom_command(
-    OUTPUT ${GEOS_GTFV3_INTERFACE_LIBRARY}
+    OUTPUT ${PYFV3_INTERFACE_LIBRARY}
     # Note below is essentially:
     #  mpirun -np 1 python file
     # but we use the CMake options as much as we can for flexibility
-    COMMAND ${CMAKE_COMMAND} -E copy_if_different ${GEOS_GTFV3_INTERFACE_FLAG_HEADER_FILE} ${CMAKE_CURRENT_BINARY_DIR}
-    COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 1 ${Python3_EXECUTABLE} ${GEOS_GTFV3_INTERFACE_SRCS}
-    BYPRODUCTS ${GEOS_GTFV3_INTERFACE_HEADER_FILE}
+    COMMAND ${CMAKE_COMMAND} -E copy_if_different ${PYFV3_INTERFACE_FLAG_HEADER_FILE} ${CMAKE_CURRENT_BINARY_DIR}
+    COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 1 ${Python3_EXECUTABLE} ${PYFV3_INTERFACE_SRCS}
+    BYPRODUCTS ${PYFV3_INTERFACE_HEADER_FILE}
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-    MAIN_DEPENDENCY ${GEOS_GTFV3_INTERFACE_SRCS}
-    COMMENT "Building gtfv3 interface library with Python"
+    MAIN_DEPENDENCY ${PYFV3_INTERFACE_SRCS}
+    COMMENT "Building pyFV3 interface library with Python"
     VERBATIM
     )
 
   # This creates a target we can use for dependencies and post build
-  add_custom_target(generate_python_interface_library DEPENDS ${GEOS_GTFV3_INTERFACE_LIBRARY})
+  add_custom_target(generate_python_interface_library DEPENDS ${PYFV3_INTERFACE_LIBRARY})
 
   # Because of the weird hacking of INTERFACE libraries below, we cannot
   # use the "usual" CMake calls to install() the .so. I think it's because
@@ -75,31 +75,31 @@ if (BUILD_GEOS_GTFV3_INTERFACE)
     # of a race condition (if install/lib/ isn't mkdir'd first)
     COMMAND ${CMAKE_COMMAND} -E make_directory ${CMAKE_INSTALL_PREFIX}/lib
     # Now we copy the file (if different...though not sure if this is useful)
-    COMMAND ${CMAKE_COMMAND} -E copy_if_different "${GEOS_GTFV3_INTERFACE_LIBRARY}" ${CMAKE_INSTALL_PREFIX}/lib
+    COMMAND ${CMAKE_COMMAND} -E copy_if_different "${PYFV3_INTERFACE_LIBRARY}" ${CMAKE_INSTALL_PREFIX}/lib
     )
 
   # We use INTERFACE libraries to create a sort of "fake" target library we can use
   # to make libFVdycoreCubed_GridComp.a depend on. It seems to work!
-  add_library(geos_gtfv3_interface_py INTERFACE)
+  add_library(pyfv3_interface_py INTERFACE)
 
   # The target_include_directories bits were essentially stolen from the esma_add_library
   # code...
-  target_include_directories(geos_gtfv3_interface_py INTERFACE
+  target_include_directories(pyfv3_interface_py INTERFACE
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
     $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}> # stubs
     # modules and copied *.h, *.inc
     $<BUILD_INTERFACE:${esma_include}/${this}>
     $<INSTALL_INTERFACE:include/${this}>
     )
-  target_link_libraries(geos_gtfv3_interface_py INTERFACE ${GEOS_GTFV3_INTERFACE_LIBRARY})
+  target_link_libraries(pyfv3_interface_py INTERFACE ${PYFV3_INTERFACE_LIBRARY})
 
   # This makes sure the library is built first
-  add_dependencies(geos_gtfv3_interface_py generate_python_interface_library)
+  add_dependencies(pyfv3_interface_py generate_python_interface_library)
 
   # This bit is to resolve an issue and Google told me to do this. I'm not
   # sure that the LIBRARY DESTINATION bit actually does anything since
   # this is using INTERFACE
-  install(TARGETS geos_gtfv3_interface_py
+  install(TARGETS pyfv3_interface_py
     EXPORT ${PROJECT_NAME}-targets
     LIBRARY DESTINATION ${CMAKE_INSTALL_PREFIX}/lib
     )
@@ -108,13 +108,13 @@ endif ()
 
 set(dependencies MAPL GFTL_SHARED::gftl-shared GMAO_hermes GEOS_Shared esmf OpenMP::OpenMP_Fortran)
 
-if (BUILD_GEOS_GTFV3_INTERFACE)
+if (BUILD_PYFV3_INTERFACE)
   esma_add_library (${this}
     SRCS ${srcs}
     SUBCOMPONENTS fvdycore
     DEPENDENCIES ${dependencies}
     DEPENDENCIES ${GFDL}
-    DEPENDENCIES geos_gtfv3_interface_py) # Make the main library depend on the Python library
+    DEPENDENCIES pyfv3_interface_py) # Make the main library depend on the Python library
 else ()
   esma_add_library (${this}
     SRCS ${srcs}
