--- python/interface/README.md
@@ -1,6 +1,0 @@
-GEOS - GTFV3 bridge
-===================
-
-Input control to GTFV3 via environement variables
-
-* `GTFV3_BACKEND`: backend for `gtFV3`, defaults to `gt:gpu`
